refactor(library): rename import_commits to import_lib_commits
